task 7 completed

+++ app/services/category.py
@@ -1,12 +1,14 @@
 from sqlalchemy.orm import Session
 
-from app.models.category import CategoryORM
 from app.repositories.category import CategoryRepository
-from app.schemas.category import CategoryRead, CategoryCreate, CategoryUpdate
+from app.schemas.category import CategoryCreate, CategoryRead, CategoryUpdate
+
+
+class CategoryNotFoundError(Exception):
+    pass
 
 
 class CategoryService:
-
     def __init__(self, db: Session):
         self.db = db
         self.repository = CategoryRepository(db)
@@ -20,18 +22,23 @@ class CategoryService:
         self.db.commit()
         return CategoryRead.model_validate(category)
 
-    def update_category(self, category_id: str, payload: CategoryUpdate) -> CategoryRead:
+    def update_category(
+        self, category_id: str, payload: CategoryUpdate
+    ) -> CategoryRead:
         category = self.repository.get_by_id(category_id)
-
-        if payload.name is not None:
-            category.name = payload.name
+        if category is not None:
+            if payload.name is not None:
+                category.name = payload.name
+        else:
+            raise CategoryNotFoundError()
 
         self.db.commit()
         return CategoryRead.model_validate(category)
 
     def delete_category(self, category_id: str) -> None:
         category = self.repository.get_by_id(category_id)
-
-        self.repository.delete(category)
+        if category is not None:
+            self.repository.delete(category)
+        else:
+            raise CategoryNotFoundError()
         self.db.commit()
-
